Move the options related code into a separate file.
- Add a 'Data' class.
- Update the code to use this new class.
