+++ source/data.py
@@ -1,0 +1,101 @@
+import json
+from pathlib import Path
+
+from PySide2.QtGui import QColor
+
+CONFIG_PATH = Path.home() / '.config' / 'screen_ruler' / 'config.json'
+
+
+class Data():
+    def __init__(self):
+        self.options = {
+            'units': 'px',  # px / cm / inch
+            'always_above': False,
+            'horizontal_orientation': True,
+            'background_color': QColor(222, 212, 33, 210),
+            'lines_color': QColor(0, 0, 0, 255),
+            'divisions_color': QColor(0, 0, 255, 255),
+            'ruler_width': 500,
+            'ruler_height': 50,
+            'options_opened': False,
+            'division_lines': False,
+        }
+        self.load()
+
+    def get(self, key):
+        return self.options[key]
+
+    def update(self, key, value):
+        self.options[key] = value
+
+    def load(self):
+        """
+            Load the options from the configuration file.
+        """
+        try:
+            text = CONFIG_PATH.read_text(encoding='utf-8')
+            options = json.loads(text)
+
+        except (FileNotFoundError, ValueError):
+            return
+
+        for key, value in options.items():
+            self.options[key] = value
+
+            # deal with the colors (init. the QColor() from)
+        backgroundColor = self.options['background_color']
+        linesColor = self.options['lines_color']
+        divisionsColor = self.options['divisions_color']
+
+        if isinstance(backgroundColor, dict):
+            self.options['background_color'] = QColor(
+                backgroundColor['red'], backgroundColor['green'],
+                backgroundColor['blue'], backgroundColor['alpha'])
+
+        if isinstance(linesColor, dict):
+            self.options['lines_color'] = QColor(
+                linesColor['red'], linesColor['green'],
+                linesColor['blue'], linesColor['alpha'])
+
+        if isinstance(divisionsColor, dict):
+            self.options['divisions_color'] = QColor(
+                divisionsColor['red'], divisionsColor['green'],
+                divisionsColor['blue'], divisionsColor['alpha'])
+
+    def save(self, currentState):
+        """
+            Write the current options values to the configuration file.
+        """
+        # need to get the individual rgba() from the QColor so that we can serialize it into json
+        background = self.options['background_color']
+        lines = self.options['lines_color']
+        divisions = self.options['divisions_color']
+
+        self.options['background_color'] = {
+            'red': background.red(),
+            'green': background.green(),
+            'blue': background.blue(),
+            'alpha': background.alpha()
+        }
+        self.options['lines_color'] = {
+            'red': lines.red(),
+            'green': lines.green(),
+            'blue': lines.blue(),
+            'alpha': lines.alpha()
+        }
+        self.options['divisions_color'] = {
+            'red': divisions.red(),
+            'green': divisions.green(),
+            'blue': divisions.blue(),
+            'alpha': divisions.alpha()
+        }
+
+        self.options['ruler_width'] = currentState['width']
+        self.options['ruler_height'] = currentState['height']
+        self.options['options_opened'] = currentState['optionsOpened']
+
+        # make sure the directory exists before trying to create the configuration file
+        if not CONFIG_PATH.parent.exists():
+            CONFIG_PATH.parent.mkdir(parents=True, exist_ok=True)
+
+        CONFIG_PATH.write_text(json.dumps(self.options), encoding='utf-8')
